traffic_gatherer/requirements.txt
fastapi
uvicorn
cicflowmeter
requests
pandas
slowapi
python-dotenv